remove auth-jwt module and move its contents to subscriptions-impl
