--- auth-jwt/readme.md
@@ -1,9 +1,0 @@
-# Feature Name
-
-This module provides API for parsing and validating JWT (JWS) tokens issued by Auth API v2.
-
-## Who can help you better understand this feature?
-- Łukasz Macionczyk
-
-## More information
-N/A
