+++ subscriptions/subscriptions-impl/src/test/java/com/duckduckgo/subscriptions/impl/auth2/AuthJwtValidatorImplTest.kt
@@ -1,7 +1,22 @@
-package com.duckduckgo.authjwt.impl
+/*
+ * Copyright (c) 2024 DuckDuckGo
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.duckduckgo.subscriptions.impl.auth2
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.duckduckgo.authjwt.api.Entitlement
 import com.duckduckgo.common.utils.CurrentTimeProvider
 import java.time.Instant
 import java.time.LocalDateTime
